xuwang

+++ cpp_module_07/ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: xuwang <xuwang@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/25 12:31:14 by xuwang            #+#    #+#             */
-/*   Updated: 2021/11/28 13:39:43 by xuwang           ###   ########.fr       */
+/*   Updated: 2021/12/06 19:38:30 by xuwang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 template <typename T>
 void iter(T *arry, size_t size, void(*f)(T const & a)) {
-    for(int i = 0; i < size; i++) {
+    for(size_t i = 0; i < size; i++) {
         (*f)(arry[i]);
     }
 };
